requirements.txt
arrow==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
langdetect==1.0.9
loging==0.0.1
loguru~=0.5.3
mysql-connector-python==8.0.27
protobuf==3.19.4
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-telegram-bot-calendar==1.0.5
requests==2.27.1
six==1.16.0
urllib3==1.26.8
win32-setctime==1.1.0
